add optional markdown support fields

+++ lib/api/users.dart
@@ -121,8 +121,7 @@ class GeneratedUserReportData implements DocumentData {
     createdAt = readDataFromJson("createdAt", json);
 
     if (json["usedSettings"] != null) {
-      usedSettings = GenerateUserReportData()
-        ..constructFromJson(json["usedSettings"] as Map<String, dynamic>);
+      usedSettings = GenerateUserReportData()..constructFromJson(json["usedSettings"] as Map<String, dynamic>);
     }
   }
 
@@ -148,18 +147,15 @@ class GenerateUserReportData implements DocumentData {
   @override
   constructFromJson(Map<String, dynamic> json) {
     if (json["frontHistory"] != null) {
-      frontHistory = GenerateUserReportDataFh()
-        ..constructFromJson(json["frontHistory"] as Map<String, dynamic>);
+      frontHistory = GenerateUserReportDataFh()..constructFromJson(json["frontHistory"] as Map<String, dynamic>);
     }
 
     if (json["customFronts"] != null) {
-      customFronts = GenerateUserReportDataCf()
-        ..constructFromJson(json["customFronts"] as Map<String, dynamic>);
+      customFronts = GenerateUserReportDataCf()..constructFromJson(json["customFronts"] as Map<String, dynamic>);
     }
 
     if (json["members"] != null) {
-      members = GenerateUserReportDataFMem()
-        ..constructFromJson(json["members"] as Map<String, dynamic>);
+      members = GenerateUserReportDataFMem()..constructFromJson(json["members"] as Map<String, dynamic>);
     }
   }
 
@@ -186,6 +182,7 @@ class UserData implements DocumentData {
   String? color;
   Map<String, UserFieldData>? fields;
   bool? patron;
+  bool? supportDescMarkdown;
 
   @override
   Map<String, dynamic> toJson() {
@@ -198,6 +195,7 @@ class UserData implements DocumentData {
     insertData("avatarUuid", avatarUuid, payload);
     insertData("avatarUrl", avatarUrl, payload);
     insertData("color", color, payload);
+    insertData("supportDescMarkdown", supportDescMarkdown, payload);
 
     // Only insert patron for cache reasons
     insertData("patron", patron, payload);
@@ -216,6 +214,7 @@ class UserData implements DocumentData {
     avatarUrl = readDataFromJson("avatarUrl", json);
     color = readDataFromJson("color", json);
     patron = readDataFromJson("patron", json);
+    supportDescMarkdown = readDataFromJson("supportDescMarkdown", json);
 
     fields = {};
 
@@ -246,12 +245,7 @@ class Users extends Collection<UserData> {
 
   @override
   Future<Document<UserData>> get(String id) async {
-    return getSimpleDocument(
-        id,
-        "v1/user",
-        "users",
-        (DocumentResponse data) => UserData()..constructFromJson(data.content),
-        () => UserData());
+    return getSimpleDocument(id, "v1/user", "users", (DocumentResponse data) => UserData()..constructFromJson(data.content), () => UserData());
   }
 
   @deprecated
@@ -269,29 +263,19 @@ class Users extends Collection<UserData> {
 
   Future<RequestResponse> setUsername(String newUsername, String userId) async {
     try {
-      var response = await SimplyHttpClient().patch(
-          Uri.parse(
-              API().connection().getRequestUrl("v1/user/username/$userId", "")),
-          body: jsonEncode({"username": newUsername}));
+      var response = await SimplyHttpClient().patch(Uri.parse(API().connection().getRequestUrl("v1/user/username/$userId", "")), body: jsonEncode({"username": newUsername}));
 
       return createResponseObject(response);
     } catch (e) {}
     return createFailResponseObject();
   }
 
-  Future<RequestResponse> generateUserReport(
-      GenerateUserReportData data) async {
-    if (data.customFronts == null &&
-        data.frontHistory == null &&
-        data.members == null) {
-      return RequestResponse(
-          false, "You must specify at least one generation type");
+  Future<RequestResponse> generateUserReport(GenerateUserReportData data) async {
+    if (data.customFronts == null && data.frontHistory == null && data.members == null) {
+      return RequestResponse(false, "You must specify at least one generation type");
     }
     try {
-      var response = await SimplyHttpClient().post(
-          Uri.parse(
-              API().connection().getRequestUrl("v1/user/generateReport", "")),
-          body: jsonEncode(data.toJson()));
+      var response = await SimplyHttpClient().post(Uri.parse(API().connection().getRequestUrl("v1/user/generateReport", "")), body: jsonEncode(data.toJson()));
       return createResponseObject(response);
     } catch (e) {}
     return createFailResponseObject();
@@ -299,11 +283,7 @@ class Users extends Collection<UserData> {
 
   Future<RequestResponse> deleteAccount() async {
     try {
-      var response = await SimplyHttpClient().delete(
-          Uri.parse(API()
-              .connection()
-              .getRequestUrl("v1/user/${API().auth().getUid()}", "")),
-          body: jsonEncode({"performDelete": true}));
+      var response = await SimplyHttpClient().delete(Uri.parse(API().connection().getRequestUrl("v1/user/${API().auth().getUid()}", "")), body: jsonEncode({"performDelete": true}));
       return createResponseObject(response);
     } catch (e) {}
     return createFailResponseObject();
@@ -311,10 +291,7 @@ class Users extends Collection<UserData> {
 
   Future<RequestResponse> deleteUserReport(String reportId) async {
     try {
-      var response = await SimplyHttpClient().delete(Uri.parse(API()
-          .connection()
-          .getRequestUrl(
-              "v1/user/${API().auth().getUid()}/report/$reportId", "")));
+      var response = await SimplyHttpClient().delete(Uri.parse(API().connection().getRequestUrl("v1/user/${API().auth().getUid()}/report/$reportId", "")));
       return createResponseObject(response);
     } catch (e) {}
     return createFailResponseObject();
@@ -322,9 +299,7 @@ class Users extends Collection<UserData> {
 
   Future<RequestResponse> exportData() async {
     try {
-      var response = await SimplyHttpClient().post(Uri.parse(API()
-          .connection()
-          .getRequestUrl("v1/user/${API().auth().getUid()}/export", "")));
+      var response = await SimplyHttpClient().post(Uri.parse(API().connection().getRequestUrl("v1/user/${API().auth().getUid()}/export", "")));
       return createResponseObject(response);
     } catch (e) {}
     return createFailResponseObject();
